+++ src/OneBlinkFormBase.tsx
@@ -48,6 +48,7 @@ import useAuth from './hooks/useAuth'
 import { formElementsService } from '@oneblink/sdk-core'
 import { TaskContext } from './hooks/useTaskContext'
 import { OnUploadAttachmentContext } from './hooks/useOnUploadAttachment'
+import { injectOptionsAcrossAllElements } from './services/injectableOptions'
 
 export type OneBlinkReadOnlyFormProps = {
   /**
@@ -186,7 +187,7 @@ function OneBlinkFormBase({
   onUploadAttachment,
 }: Props) {
   const isOffline = useIsOffline()
-  const { isUsingFormsKey } = useAuth()
+  const { isUsingFormsKey, userProfile } = useAuth()
 
   const theme = React.useMemo(
     () =>
@@ -209,6 +210,14 @@ function OneBlinkFormBase({
     }
     return formElementsService.determineIsInfoPage(definition)
   }, [definition, isInfoPageProp])
+
+  const taskContextValue = React.useMemo(() => {
+    return {
+      task,
+      taskGroup,
+      taskGroupInstance,
+    }
+  }, [task, taskGroup, taskGroupInstance])
 
   //
   //
@@ -589,8 +598,18 @@ function OneBlinkFormBase({
 
       allowNavigation()
 
+      // transplant injected options on the definition
+      const elementsWithInjectedOptions = injectOptionsAcrossAllElements({
+        elements: definition.elements,
+        submission: submissionData.submission,
+        taskContext: taskContextValue,
+        userProfile: userProfile ?? undefined,
+      })
       onSubmit({
-        definition,
+        definition: {
+          ...definition,
+          elements: elementsWithInjectedOptions,
+        },
         submission: submissionData.submission,
         captchaTokens: submissionData.captchaTokens,
       })
@@ -602,13 +621,15 @@ function OneBlinkFormBase({
       checkBsbAreValidating,
       formElementsValidation,
       checkBsbsCanBeSubmitted,
+      checkAttachmentsCanBeSubmitted,
       definition,
       attachmentRetentionInDays,
       isOffline,
       isPendingQueueEnabled,
       allowNavigation,
+      taskContextValue,
+      userProfile,
       onSubmit,
-      checkAttachmentsCanBeSubmitted,
       setFormSubmission,
     ],
   )
@@ -754,14 +775,6 @@ function OneBlinkFormBase({
   // #endregion
   //
   //
-
-  const taskContextValue = React.useMemo(() => {
-    return {
-      task,
-      taskGroup,
-      taskGroupInstance,
-    }
-  }, [task, taskGroup, taskGroupInstance])
 
   const lastElementUpdatedExistsOnForm = React.useMemo(() => {
     return !!formElementsService.findFormElement(