+++ src/hooks/useFormElementOptions.ts
@@ -6,6 +6,10 @@ import {
   UpdateFormElementsHandler,
 } from '../types/form'
 import { useLoadDynamicOptionsEffect } from './useDynamicOptionsLoaderState'
+import useFormSubmissionModel from './useFormSubmissionModelContext'
+import processInjectableOption from '../services/injectableOptions'
+import useTaskContext from './useTaskContext'
+import useAuth from './useAuth'
 
 export default function useFormElementOptions<T>({
   element,
@@ -24,6 +28,10 @@ export default function useFormElementOptions<T>({
   onFilter?: (choiceElementOption: FormTypes.ChoiceElementOption) => boolean
   onUpdateFormElements: UpdateFormElementsHandler
 }) {
+  const taskContext = useTaskContext()
+  const { userProfile } = useAuth()
+
+  const { formSubmissionModel, elements } = useFormSubmissionModel()
   const conditionallyShownOptions = conditionallyShownOptionsElement?.options
   //options that are shown due to conditional logic
   const shownOptions = React.useMemo<FormTypes.ChoiceElementOption[]>(() => {
@@ -41,19 +49,38 @@ export default function useFormElementOptions<T>({
     })
   }, [conditionallyShownOptions, element.options, onFilter])
 
+  //
+  const withInjectedOptions = React.useMemo(() => {
+    return shownOptions.reduce<FormTypes.ChoiceElementOption[]>(
+      (memo, option) => {
+        const newOptions = processInjectableOption({
+          option,
+          submission: formSubmissionModel,
+          formElements: elements,
+          taskContext,
+          userProfile: userProfile ?? undefined,
+        })
+        memo.push(...newOptions)
+
+        return memo
+      },
+      [],
+    )
+  }, [elements, formSubmissionModel, shownOptions, taskContext, userProfile])
+
   //options that are shown based on conditional logic and user input
   const filteredOptions = React.useMemo<FormTypes.ChoiceElementOption[]>(() => {
-    const reducedOptions = shownOptions.filter(
+    const reducedOptions = withInjectedOptions.filter(
       (option) => !onFilter || (onFilter(option) && !option.displayAlways),
     )
     if (element.type === 'autocomplete') {
-      const alwaysShownOptions = shownOptions.filter(
+      const alwaysShownOptions = withInjectedOptions.filter(
         (option) => option.displayAlways,
       )
       reducedOptions.push(...alwaysShownOptions)
     }
     return reducedOptions
-  }, [shownOptions, element.type, onFilter])
+  }, [withInjectedOptions, element.type, onFilter])
 
   React.useEffect(() => {
     if (
@@ -66,7 +93,7 @@ export default function useFormElementOptions<T>({
     if (
       typeof value === 'string' &&
       value &&
-      !shownOptions.some((option) => value === option.value)
+      !withInjectedOptions.some((option) => value === option.value)
     ) {
       onChange(element, {
         value: undefined,
@@ -76,7 +103,7 @@ export default function useFormElementOptions<T>({
 
     if (Array.isArray(value)) {
       const newValue = value.filter((selectedValue) =>
-        shownOptions.some((option) => selectedValue === option.value),
+        withInjectedOptions.some((option) => selectedValue === option.value),
       )
       if (newValue.length !== value.length) {
         const newValueArray = newValue.length ? newValue : undefined
@@ -91,6 +118,7 @@ export default function useFormElementOptions<T>({
     onChange,
     value,
     conditionallyShownOptionsElement?.dependencyIsLoading,
+    withInjectedOptions,
   ])
 
   useLoadDynamicOptionsEffect(element, onUpdateFormElements)